backend/requirements.txt
imutils==0.5.3
scipy==1.3.1
numpy==1.17.1
opencv-python==3.4.5.20
python-magic==0.4.18
waitress==1.4.4
Flask==1.1.2
